# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Raffle Draw",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/agent/all-stakes": {
            "get": {
                "tags": [
                    "Agent"
                ],
                "summary": "Get Machines",
                "operationId": "65b36376ccb9ace8f3e74fc2927155e9",
                "parameters": [
                    {
                        "name": "agent_id",
                        "in": "query",
                        "description": "id of the agent",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get category id",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/agent/get-raffle": {
            "get": {
                "tags": [
                    "Agent"
                ],
                "summary": "Get Withdrawal",
                "operationId": "c0926e05316be9bc91edfa3ba99a0205",
                "parameters": [
                    {
                        "name": "agent_id",
                        "in": "query",
                        "description": "id of the user",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Getting withdrawal Sucessful",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "User Not Found",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "validation Error",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}